Alert. Use the Alert dialog to show alerts to the users. In particular use the alert to confirm the deletion of favorites. *  Use alert dialogs to show alerts to users. *  Take an action upon user confirmation.

+++ src/components/FavoriteComponent.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
-import swipeout from 'react-native-swipeout';
+import Swipeout from 'react-native-swipeout';
 
 import { baseUrl } from '../shared/baseUrl';
 import { Loading } from './LoadingComponent';
 import { deleteFavorite } from '../redux/ActionCreators';
-import Swipeout from 'react-native-swipeout';
+
 
 const mapStateToProps = state => {
     return {
@@ -35,7 +35,24 @@ class Favorites extends Component {
                 {
                     text: 'Delete',
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorrite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                {
+                                    text: 'Cancel',
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
